Pull request #119: Feature/MGMIDN-1228 create api to sort the courses
Merge in DA-NANG-INTERNSHIP/course-for-you from feature/MGMIDN-1228-create-api-to-sort-the-courses to feature/MGMIDN-1105-list-of-courses-can-be-sorted-by-criteria

Squashed commit of the following:

commit 56041c4a1e3811518266f4f9d2bc69b8f2b7d5d5
Author: Lighttx26 <ncnhatquang26@gmail.com>
Date:   Thu Jul 11 08:51:50 2024 +0700

    MGMIDN-1228: Remove method count available courses in CourseService

commit a5894b70ced6fbfaf6e0ada7e94586ca5afe12fd
Author: Lighttx26 <ncnhatquang26@gmail.com>
Date:   Thu Jul 11 00:03:10 2024 +0700

    MGMIDN-1228: Remove unused import

commit f57e5333ec5709ea32fdf3635adaa5dff1de7a1c
Author: Lighttx26 <ncnhatquang26@gmail.com>
Date:   Thu Jul 11 00:02:21 2024 +0700

    MGMIDN-1228: Fix wrong totalElements of pageable result
    - Use subquery instead of join in sortByEnrollmentCountDesc Specification
    - Update build.gradle to exclude specification package

commit 568d39d237fdfffef41aa15d84a5864edb7a5886
Author: Lighttx26 <ncnhatquang26@gmail.com>
Date:   Wed Jul 10 18:09:55 2024 +0700

    MGMIDN-1228: Use Specification to sort available courses
    - Use count query in another methods

commit db4753807e128b5c155e31fdb7184012075c2584
Author: Lighttx26 <ncnhatquang26@gmail.com>
Date:   Wed Jul 10 09:31:47 2024 +0700

    MGMIDN-1227: Query to sort available courses by created date & enrollments

+++ server/src/main/java/com/mgmtp/cfu/dto/CoursePageDTO.java
@@ -12,5 +12,5 @@ import java.util.List;
 public class CoursePageDTO {
     private List<CourseOverviewDTO> courses;
     private int totalPages;
-    private long totalElements;
+    private int totalElements;
 }
